[Reporting] Add support for running the reporting infrastructure locally
This patch adds support for running the reporting infrastructure
locally, eg. during development and maintenance of the client runners.

The `docker-compose.yml` file allows to launch an Elasticsearch cluster
and a Kibana instance with reporting dashboards pre-loaded.

The `Makefile` facilitates the setup process: starting the stack,
loading the neccessary index templates, the transform definition,
and the Kibana dashboard and index patterns.