[UI] Add the web frontend
This patch adds the assets for the simple web frontend,
displaying the clients benchmarks data.

The `public/index.html` file contains the UI itself, implemented in D3.js.

The `api/index.js` contains the HTTP API, implemented in Node.js and deployed as a Google Cloud Function.

Run `make` to see a list of available targets, including ones for running the UI locally.